fixing the inconsistency due to removed perforations
for multisegment wells. Before this the multisegment wells are built
based on the Well information from parser. Some perforations can removed
during inactive cells. This commit is trying to fix this.

+++ opm/autodiff/WellInterface.hpp
@@ -125,7 +125,7 @@ namespace Opm
         virtual void init(const PhaseUsage* phase_usage_arg,
                           const std::vector<double>& depth_arg,
                           const double gravity_arg,
-                          const int num_cells);
+                          const size_t num_cells);
 
         virtual void initPrimaryVariablesEvaluation() const = 0;
